+++ src/include/memory_stats.hpp
@@ -4,9 +4,6 @@
 
 namespace duckdb {
 
-// Forward declaration.
-class ClientContext;
-
 struct MemoryInfo {
 	uint64_t total_memory = 0;
 	uint64_t used_memory = 0;
@@ -18,6 +15,6 @@ struct MemoryInfo {
 };
 
 // Get memory information for the current platform
-MemoryInfo GetMemoryInfo(ClientContext &context);
+MemoryInfo GetMemoryInfo();
 
 } // namespace duckdb