+++ src/include/os_info.hpp
@@ -5,7 +5,6 @@
 #include "duckdb/common/vector.hpp"
 
 namespace duckdb {
-class ClientContext;
 
 struct OSInfo {
 	string name;
@@ -19,6 +18,6 @@ struct OSInfo {
 };
 
 // Get OS information for the current platform
-OSInfo GetOSInfo(ClientContext &context);
+OSInfo GetOSInfo();
 
 } // namespace duckdb